r/(linux|windows)_virtual_machine_scale_set: reusing code/support for spot instances
