+++ azurerm/internal/services/compute/resource_arm_windows_virtual_machine_scale_set.go
@@ -122,7 +122,7 @@ func resourceArmWindowsVirtualMachineScaleSet() *schema.Resource {
 
 			"automatic_os_upgrade_policy": VirtualMachineScaleSetAutomatedOSUpgradePolicySchema(),
 
-			"boot_diagnostics": VirtualMachineScaleSetBootDiagnosticsSchema(),
+			"boot_diagnostics": bootDiagnosticsSchema(),
 
 			"computer_name_prefix": {
 				Type:     schema.TypeString,
@@ -153,7 +153,7 @@ func resourceArmWindowsVirtualMachineScaleSet() *schema.Resource {
 			},
 
 			"eviction_policy": {
-				// only applicable when `priority` is set to `Low`
+				// only applicable when `priority` is set to `Spot`
 				Type:     schema.TypeString,
 				Optional: true,
 				ForceNew: true,
@@ -194,7 +194,7 @@ func resourceArmWindowsVirtualMachineScaleSet() *schema.Resource {
 				Default:  true,
 			},
 
-			"plan": PlanSchema(),
+			"plan": planSchema(),
 
 			"priority": {
 				Type:     schema.TypeString,
@@ -202,8 +202,8 @@ func resourceArmWindowsVirtualMachineScaleSet() *schema.Resource {
 				ForceNew: true,
 				Default:  string(compute.Regular),
 				ValidateFunc: validation.StringInSlice([]string{
-					string(compute.Low),
 					string(compute.Regular),
+					string(compute.Spot),
 				}, false),
 			},
 
@@ -225,39 +225,7 @@ func resourceArmWindowsVirtualMachineScaleSet() *schema.Resource {
 
 			"rolling_upgrade_policy": VirtualMachineScaleSetRollingUpgradePolicySchema(),
 
-			"secret": {
-				Type:     schema.TypeList,
-				Optional: true,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						// whilst this isn't present in the nested object it's required when this is specified
-						"key_vault_id": {
-							Type:         schema.TypeString,
-							Required:     true,
-							ValidateFunc: azure.ValidateResourceID,
-						},
-
-						"certificate": {
-							Type:     schema.TypeSet,
-							Required: true,
-							MinItems: 1,
-							Elem: &schema.Resource{
-								Schema: map[string]*schema.Schema{
-									"store": {
-										Type:     schema.TypeString,
-										Required: true,
-									},
-									"url": {
-										Type:         schema.TypeString,
-										Required:     true,
-										ValidateFunc: azure.ValidateKeyVaultChildId,
-									},
-								},
-							},
-						},
-					},
-				},
-			},
+			"secret": windowsSecretSchema(),
 
 			"single_placement_group": {
 				Type:     schema.TypeBool,
@@ -271,7 +239,7 @@ func resourceArmWindowsVirtualMachineScaleSet() *schema.Resource {
 				ValidateFunc: azure.ValidateResourceID,
 			},
 
-			"source_image_reference": VirtualMachineScaleSetSourceImageReferenceSchema(),
+			"source_image_reference": sourceImageReferenceSchema(),
 
 			"tags": tags.Schema(),
 
@@ -378,7 +346,7 @@ func resourceArmWindowsVirtualMachineScaleSetCreate(d *schema.ResourceData, meta
 	additionalUnattendConfig := expandWindowsVirtualMachineScaleSetAdditionalUnattendConfig(additionalUnattendConfigRaw)
 
 	bootDiagnosticsRaw := d.Get("boot_diagnostics").([]interface{})
-	bootDiagnostics := ExpandVirtualMachineScaleSetBootDiagnostics(bootDiagnosticsRaw)
+	bootDiagnostics := expandBootDiagnostics(bootDiagnosticsRaw)
 
 	dataDisksRaw := d.Get("data_disk").([]interface{})
 	dataDisks := ExpandVirtualMachineScaleSetDataDisk(dataDisksRaw)
@@ -399,11 +367,11 @@ func resourceArmWindowsVirtualMachineScaleSetCreate(d *schema.ResourceData, meta
 	osDisk := ExpandVirtualMachineScaleSetOSDisk(osDiskRaw, compute.Windows)
 
 	planRaw := d.Get("plan").([]interface{})
-	plan := ExpandPlan(planRaw)
+	plan := expandPlan(planRaw)
 
 	sourceImageReferenceRaw := d.Get("source_image_reference").([]interface{})
 	sourceImageId := d.Get("source_image_id").(string)
-	sourceImageReference, err := ExpandVirtualMachineScaleSetSourceImageReference(sourceImageReferenceRaw, sourceImageId)
+	sourceImageReference, err := expandSourceImageReference(sourceImageReferenceRaw, sourceImageId)
 	if err != nil {
 		return err
 	}
@@ -438,7 +406,7 @@ func resourceArmWindowsVirtualMachineScaleSetCreate(d *schema.ResourceData, meta
 	winRmListeners := expandWindowsVirtualMachineScaleSetWinRMListeners(winRmListenersRaw)
 
 	secretsRaw := d.Get("secret").([]interface{})
-	secrets := expandWindowsVirtualMachineScaleSetSecrets(secretsRaw)
+	secrets := expandWindowsSecrets(secretsRaw)
 
 	zonesRaw := d.Get("zones").([]interface{})
 	zones := azure.ExpandZones(zonesRaw)
@@ -495,8 +463,8 @@ func resourceArmWindowsVirtualMachineScaleSetCreate(d *schema.ResourceData, meta
 	}
 
 	if v, ok := d.Get("max_bid_price").(float64); ok && v > 0 {
-		if priority != compute.Low {
-			return fmt.Errorf("`max_bid_price` can only be configured when `priority` is set to `Low`")
+		if priority != compute.Spot {
+			return fmt.Errorf("`max_bid_price` can only be configured when `priority` is set to `Spot`")
 		}
 
 		virtualMachineProfile.BillingProfile = &compute.BillingProfile{
@@ -509,12 +477,12 @@ func resourceArmWindowsVirtualMachineScaleSetCreate(d *schema.ResourceData, meta
 	}
 
 	if evictionPolicyRaw, ok := d.GetOk("eviction_policy"); ok {
-		if virtualMachineProfile.Priority != compute.Low {
-			return fmt.Errorf("An `eviction_policy` can only be specified when `priority` is set to `low`")
+		if virtualMachineProfile.Priority != compute.Spot {
+			return fmt.Errorf("An `eviction_policy` can only be specified when `priority` is set to `Spot`")
 		}
 		virtualMachineProfile.EvictionPolicy = compute.VirtualMachineEvictionPolicyTypes(evictionPolicyRaw.(string))
-	} else if priority == compute.Low {
-		return fmt.Errorf("An `eviction_policy` must be specified when `priority` is set to `low`")
+	} else if priority == compute.Spot {
+		return fmt.Errorf("An `eviction_policy` must be specified when `priority` is set to `Spot`")
 	}
 
 	if len(additionalUnattendConfigRaw) > 0 {
@@ -642,8 +610,8 @@ func resourceArmWindowsVirtualMachineScaleSetUpdate(d *schema.ResourceData, meta
 
 	priority := compute.VirtualMachinePriorityTypes(d.Get("priority").(string))
 	if d.HasChange("max_bid_price") {
-		if priority != compute.Low {
-			return fmt.Errorf("`max_bid_price` can only be configured when `priority` is set to `Low`")
+		if priority != compute.Spot {
+			return fmt.Errorf("`max_bid_price` can only be configured when `priority` is set to `Spot`")
 		}
 
 		updateProps.VirtualMachineProfile.BillingProfile = &compute.BillingProfile{
@@ -696,7 +664,7 @@ func resourceArmWindowsVirtualMachineScaleSetUpdate(d *schema.ResourceData, meta
 
 		if d.HasChange("secret") {
 			secretsRaw := d.Get("secret").([]interface{})
-			osProfile.Secrets = expandWindowsVirtualMachineScaleSetSecrets(secretsRaw)
+			osProfile.Secrets = expandWindowsSecrets(secretsRaw)
 		}
 
 		updateProps.VirtualMachineProfile.OsProfile = &osProfile
@@ -720,7 +688,7 @@ func resourceArmWindowsVirtualMachineScaleSetUpdate(d *schema.ResourceData, meta
 		if d.HasChange("source_image_id") || d.HasChange("source_image_reference") {
 			sourceImageReferenceRaw := d.Get("source_image_reference").([]interface{})
 			sourceImageId := d.Get("source_image_id").(string)
-			sourceImageReference, err := ExpandVirtualMachineScaleSetSourceImageReference(sourceImageReferenceRaw, sourceImageId)
+			sourceImageReference, err := expandSourceImageReference(sourceImageReferenceRaw, sourceImageId)
 			if err != nil {
 				return err
 			}
@@ -750,7 +718,7 @@ func resourceArmWindowsVirtualMachineScaleSetUpdate(d *schema.ResourceData, meta
 		updateInstances = true
 
 		bootDiagnosticsRaw := d.Get("boot_diagnostics").([]interface{})
-		updateProps.VirtualMachineProfile.DiagnosticsProfile = ExpandVirtualMachineScaleSetBootDiagnostics(bootDiagnosticsRaw)
+		updateProps.VirtualMachineProfile.DiagnosticsProfile = expandBootDiagnostics(bootDiagnosticsRaw)
 	}
 
 	if d.HasChange("identity") {
@@ -765,7 +733,7 @@ func resourceArmWindowsVirtualMachineScaleSetUpdate(d *schema.ResourceData, meta
 
 	if d.HasChange("plan") {
 		planRaw := d.Get("plan").([]interface{})
-		update.Plan = ExpandPlan(planRaw)
+		update.Plan = expandPlan(planRaw)
 	}
 
 	if d.HasChange("sku") || d.HasChange("instances") {
@@ -916,7 +884,7 @@ func resourceArmWindowsVirtualMachineScaleSetRead(d *schema.ResourceData, meta i
 		return fmt.Errorf("Error setting `identity`: %+v", err)
 	}
 
-	if err := d.Set("plan", FlattenPlan(resp.Plan)); err != nil {
+	if err := d.Set("plan", flattenPlan(resp.Plan)); err != nil {
 		return fmt.Errorf("Error setting `plan`: %+v", err)
 	}
 
@@ -957,7 +925,7 @@ func resourceArmWindowsVirtualMachineScaleSetRead(d *schema.ResourceData, meta i
 	}
 
 	if profile := props.VirtualMachineProfile; profile != nil {
-		if err := d.Set("boot_diagnostics", FlattenVirtualMachineScaleSetBootDiagnostics(profile.DiagnosticsProfile)); err != nil {
+		if err := d.Set("boot_diagnostics", flattenBootDiagnostics(profile.DiagnosticsProfile)); err != nil {
 			return fmt.Errorf("Error setting `boot_diagnostics`: %+v", err)
 		}
 
@@ -981,7 +949,7 @@ func resourceArmWindowsVirtualMachineScaleSetRead(d *schema.ResourceData, meta i
 				return fmt.Errorf("Error setting `data_disk`: %+v", err)
 			}
 
-			if err := d.Set("source_image_reference", FlattenVirtualMachineScaleSetSourceImageReference(storageProfile.ImageReference)); err != nil {
+			if err := d.Set("source_image_reference", flattenSourceImageReference(storageProfile.ImageReference)); err != nil {
 				return fmt.Errorf("Error setting `source_image_reference`: %+v", err)
 			}
 
@@ -997,7 +965,7 @@ func resourceArmWindowsVirtualMachineScaleSetRead(d *schema.ResourceData, meta i
 			d.Set("admin_username", osProfile.AdminUsername)
 			d.Set("computer_name_prefix", osProfile.ComputerNamePrefix)
 
-			if err := d.Set("secret", flattenWindowsVirtualMachineScaleSetSecrets(osProfile.Secrets)); err != nil {
+			if err := d.Set("secret", flattenWindowsSecrets(osProfile.Secrets)); err != nil {
 				return fmt.Errorf("Error setting `secret`: %+v", err)
 			}
 
@@ -1157,80 +1125,6 @@ func flattenWindowsVirtualMachineScaleSetAdditionalUnattendConfig(input *[]compu
 		output = append(output, map[string]interface{}{
 			"content": content,
 			"setting": string(v.SettingName),
-		})
-	}
-
-	return output
-}
-
-func expandWindowsVirtualMachineScaleSetSecrets(input []interface{}) *[]compute.VaultSecretGroup {
-	output := make([]compute.VaultSecretGroup, 0)
-
-	for _, raw := range input {
-		v := raw.(map[string]interface{})
-
-		keyVaultId := v["key_vault_id"].(string)
-		certificatesRaw := v["certificate"].(*schema.Set).List()
-		certificates := make([]compute.VaultCertificate, 0)
-		for _, certificateRaw := range certificatesRaw {
-			certificateV := certificateRaw.(map[string]interface{})
-
-			store := certificateV["store"].(string)
-			url := certificateV["url"].(string)
-			certificates = append(certificates, compute.VaultCertificate{
-				CertificateStore: utils.String(store),
-				CertificateURL:   utils.String(url),
-			})
-		}
-
-		output = append(output, compute.VaultSecretGroup{
-			SourceVault: &compute.SubResource{
-				ID: utils.String(keyVaultId),
-			},
-			VaultCertificates: &certificates,
-		})
-	}
-
-	return &output
-}
-
-func flattenWindowsVirtualMachineScaleSetSecrets(input *[]compute.VaultSecretGroup) []interface{} {
-	if input == nil {
-		return []interface{}{}
-	}
-
-	output := make([]interface{}, 0)
-
-	for _, v := range *input {
-		keyVaultId := ""
-		if v.SourceVault != nil && v.SourceVault.ID != nil {
-			keyVaultId = *v.SourceVault.ID
-		}
-
-		certificates := make([]interface{}, 0)
-
-		if v.VaultCertificates != nil {
-			for _, c := range *v.VaultCertificates {
-				store := ""
-				if c.CertificateStore != nil {
-					store = *c.CertificateStore
-				}
-
-				url := ""
-				if c.CertificateURL != nil {
-					url = *c.CertificateURL
-				}
-
-				certificates = append(certificates, map[string]interface{}{
-					"store": store,
-					"url":   url,
-				})
-			}
-		}
-
-		output = append(output, map[string]interface{}{
-			"key_vault_id": keyVaultId,
-			"certificate":  certificates,
 		})
 	}
