+++ azurerm/internal/services/compute/virtual_machine_scale_set.go
@@ -88,63 +88,6 @@ func FlattenVirtualMachineScaleSetAdditionalCapabilities(input *compute.Addition
 	}
 }
 
-func VirtualMachineScaleSetBootDiagnosticsSchema() *schema.Schema {
-	return &schema.Schema{
-		Type:     schema.TypeList,
-		Optional: true,
-		MaxItems: 1,
-		Elem: &schema.Resource{
-			Schema: map[string]*schema.Schema{
-				// TODO: should this be `storage_account_endpoint`?
-				"storage_account_uri": {
-					Type:     schema.TypeString,
-					Required: true,
-					// TODO: validation
-				},
-			},
-		},
-	}
-}
-
-func ExpandVirtualMachineScaleSetBootDiagnostics(input []interface{}) *compute.DiagnosticsProfile {
-	if len(input) == 0 {
-		return &compute.DiagnosticsProfile{
-			BootDiagnostics: &compute.BootDiagnostics{
-				Enabled:    utils.Bool(false),
-				StorageURI: utils.String(""),
-			},
-		}
-	}
-
-	raw := input[0].(map[string]interface{})
-
-	storageAccountURI := raw["storage_account_uri"].(string)
-
-	return &compute.DiagnosticsProfile{
-		BootDiagnostics: &compute.BootDiagnostics{
-			Enabled:    utils.Bool(true),
-			StorageURI: utils.String(storageAccountURI),
-		},
-	}
-}
-
-func FlattenVirtualMachineScaleSetBootDiagnostics(input *compute.DiagnosticsProfile) []interface{} {
-	if input == nil || input.BootDiagnostics == nil || input.BootDiagnostics.Enabled == nil || !*input.BootDiagnostics.Enabled {
-		return []interface{}{}
-	}
-
-	storageAccountUri := ""
-	if input.BootDiagnostics.StorageURI != nil {
-		storageAccountUri = *input.BootDiagnostics.StorageURI
-	}
-
-	return []interface{}{
-		map[string]interface{}{
-			"storage_account_uri": storageAccountUri,
-		},
-	}
-}
-
 func VirtualMachineScaleSetIdentitySchema() *schema.Schema {
 	return &schema.Schema{
 		Type:     schema.TypeList,
@@ -1103,90 +1046,6 @@ func FlattenVirtualMachineScaleSetOSDisk(input *compute.VirtualMachineScaleSetOS
 			"storage_account_type":      storageAccountType,
 			"write_accelerator_enabled": writeAcceleratorEnabled,
 			"disk_encryption_set_id":    diskEncryptionSetId,
-		},
-	}
-}
-
-func VirtualMachineScaleSetSourceImageReferenceSchema() *schema.Schema {
-	// whilst originally I was hoping we could use the 'id' from `azurerm_platform_image' unfortunately Azure doesn't
-	// like this as a value for the 'id' field:
-	// Id /...../Versions/16.04.201909091 is not a valid resource reference."
-	// as such the image is split into two fields (source_image_id and source_image_reference) to provide better validation
-	return &schema.Schema{
-		Type:          schema.TypeList,
-		Optional:      true,
-		MaxItems:      1,
-		ConflictsWith: []string{"source_image_id"},
-		Elem: &schema.Resource{
-			Schema: map[string]*schema.Schema{
-				"publisher": {
-					Type:     schema.TypeString,
-					Required: true,
-				},
-				"offer": {
-					Type:     schema.TypeString,
-					Required: true,
-				},
-				"sku": {
-					Type:     schema.TypeString,
-					Required: true,
-				},
-				"version": {
-					Type:     schema.TypeString,
-					Required: true,
-				},
-			},
-		},
-	}
-}
-
-func ExpandVirtualMachineScaleSetSourceImageReference(referenceInput []interface{}, imageId string) (*compute.ImageReference, error) {
-	if imageId != "" {
-		return &compute.ImageReference{
-			ID: utils.String(imageId),
-		}, nil
-	}
-
-	if len(referenceInput) == 0 {
-		return nil, fmt.Errorf("Either a `source_image_id` or a `source_image_reference` block must be specified!")
-	}
-
-	raw := referenceInput[0].(map[string]interface{})
-	return &compute.ImageReference{
-		Publisher: utils.String(raw["publisher"].(string)),
-		Offer:     utils.String(raw["offer"].(string)),
-		Sku:       utils.String(raw["sku"].(string)),
-		Version:   utils.String(raw["version"].(string)),
-	}, nil
-}
-
-func FlattenVirtualMachineScaleSetSourceImageReference(input *compute.ImageReference) []interface{} {
-	// since the image id is pulled out as a separate field, if that's set we should return an empty block here
-	if input == nil || input.ID != nil {
-		return []interface{}{}
-	}
-
-	var publisher, offer, sku, version string
-
-	if input.Publisher != nil {
-		publisher = *input.Publisher
-	}
-	if input.Offer != nil {
-		offer = *input.Offer
-	}
-	if input.Sku != nil {
-		sku = *input.Sku
-	}
-	if input.Version != nil {
-		version = *input.Version
-	}
-
-	return []interface{}{
-		map[string]interface{}{
-			"publisher": publisher,
-			"offer":     offer,
-			"sku":       sku,
-			"version":   version,
 		},
 	}
 }
