+++ azurerm/internal/services/compute/virtual_machine_scale_set_extension_test.go
@@ -1,0 +1,67 @@
+package compute
+
+import "testing"
+
+func TestParseVirtualMachineScaleSetExtensionID(t *testing.T) {
+	testData := []struct {
+		Name     string
+		Input    string
+		Expected *VirtualMachineScaleSetExtensionResourceID
+	}{
+		{
+			Name:     "Empty",
+			Input:    "",
+			Expected: nil,
+		},
+		{
+			Name:     "No Virtual Machine Scale Set Segment",
+			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo",
+			Expected: nil,
+		},
+		{
+			Name:     "No Virtual Machine Scale Set Value",
+			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/virtualMachineScaleSets/",
+			Expected: nil,
+		},
+		{
+			Name:     "No Extensions Segment",
+			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/virtualMachineScaleSets/machine1",
+			Expected: nil,
+		},
+		{
+			Name:     "No Extensions Value",
+			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/virtualMachineScaleSets/machine1/extensions/",
+			Expected: nil,
+		},
+		{
+			Name:  "Completed",
+			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/virtualMachineScaleSets/machine1/extensions/extension1",
+			Expected: &VirtualMachineScaleSetExtensionResourceID{
+				Name:               "extension1",
+				VirtualMachineName: "machine1",
+				ResourceGroup:      "foo",
+			},
+		},
+	}
+
+	for _, v := range testData {
+		t.Logf("[DEBUG] Testing %q", v.Name)
+
+		actual, err := ParseVirtualMachineScaleSetExtensionID(v.Input)
+		if err != nil {
+			if v.Expected == nil {
+				continue
+			}
+
+			t.Fatalf("Expected a value but got an error: %s", err)
+		}
+
+		if actual.Name != v.Expected.Name {
+			t.Fatalf("Expected %q but got %q for Name", v.Expected.Name, actual.Name)
+		}
+
+		if actual.ResourceGroup != v.Expected.ResourceGroup {
+			t.Fatalf("Expected %q but got %q for ResourceGroup", v.Expected.ResourceGroup, actual.ResourceGroup)
+		}
+	}
+}
