+++ azurerm/internal/services/compute/virtual_machine_scale_set_test.go
@@ -1,0 +1,56 @@
+package compute
+
+import "testing"
+
+func TestParseVirtualMachineScaleSetID(t *testing.T) {
+	testData := []struct {
+		Name     string
+		Input    string
+		Expected *VirtualMachineScaleSetResourceID
+	}{
+		{
+			Name:     "Empty",
+			Input:    "",
+			Expected: nil,
+		},
+		{
+			Name:     "No Virtual Machine Scale Set Segment",
+			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo",
+			Expected: nil,
+		},
+		{
+			Name:     "No Virtual Machine Scale Set Value",
+			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/virtualMachineScaleSets/",
+			Expected: nil,
+		},
+		{
+			Name:  "Completed",
+			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/virtualMachineScaleSets/example",
+			Expected: &VirtualMachineScaleSetResourceID{
+				Name:          "example",
+				ResourceGroup: "foo",
+			},
+		},
+	}
+
+	for _, v := range testData {
+		t.Logf("[DEBUG] Testing %q", v.Name)
+
+		actual, err := ParseVirtualMachineScaleSetID(v.Input)
+		if err != nil {
+			if v.Expected == nil {
+				continue
+			}
+
+			t.Fatalf("Expected a value but got an error: %s", err)
+		}
+
+		if actual.Name != v.Expected.Name {
+			t.Fatalf("Expected %q but got %q for Name", v.Expected.Name, actual.Name)
+		}
+
+		if actual.ResourceGroup != v.Expected.ResourceGroup {
+			t.Fatalf("Expected %q but got %q for ResourceGroup", v.Expected.ResourceGroup, actual.ResourceGroup)
+		}
+	}
+}
